moved active record types out of module
