+++ server/app/graphql/types/drug_claim_type_type.rb
@@ -2,6 +2,6 @@ module Types
   class DrugClaimTypeType < Types::BaseObject
     field :id, ID, null: false
     field :type, String, null: false
-    field :drug_claim, Types::DrugClaimType, null: false
+    field :drug_claims, [Types::DrugClaimType], null: false
   end
 end
