+++ server/app/models/interaction_claim_type.rb
